+++ examples/apps/react-todomvc-cs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # todomvc-cs
 
+## 1.2.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@9.6.0
+  - @inglorious/react-store@17.0.0
+
 ## 1.2.25
 
 ### Patch Changes