+++ examples/apps/react-todomvc-rtk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # react-todomvc-rtk
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@9.6.0
+
 ## 0.1.1
 
 ### Patch Changes