+++ examples/apps/web-charts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-charts
 
+## 0.1.2
+
+### Patch Changes
+
+- @inglorious/web@4.2.5
+- @inglorious/charts@1.0.2
+
 ## 0.1.1
 
 ### Patch Changes