+++ examples/apps/web-form/CHANGELOG.md
@@ -1,5 +1,11 @@
 # lit-todomvc
 
+## 0.1.43
+
+### Patch Changes
+
+- @inglorious/web@4.2.5
+
 ## 0.1.42
 
 ### Patch Changes