+++ examples/apps/web-todomvc-ace/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web-todomvc-ace
 
+## 0.1.12
+
+### Patch Changes
+
+- @inglorious/web@4.2.5
+
 ## 0.1.11
 
 ### Patch Changes