+++ examples/apps/web-todomvc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # lit-todomvc
 
+## 0.1.45
+
+### Patch Changes
+
+- @inglorious/web@4.2.5
+
 ## 0.1.44
 
 ### Patch Changes