+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/charts
 
+## 1.0.2
+
+### Patch Changes
+
+- @inglorious/web@4.2.5
+
 ## 1.0.1
 
 ### Patch Changes