+++ packages/motion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/motion
 
+## 0.1.1
+
+### Patch Changes
+
+- @inglorious/web@4.2.5
+
 ## 0.1.0
 
 ### Minor Changes