+++ packages/ssx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/ssx
 
+## 1.6.5
+
+### Patch Changes
+
+- @inglorious/web@4.2.5
+
 ## 1.6.4
 
 ### Patch Changes